fix: fix id names for signal and change scale factor from float to big decimal

+++ sinaleasy-back/src/main/java/com/sinalez/sinaleasy_back/controllers/SignalController.java
@@ -35,7 +35,7 @@ public class SignalController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<List<Signal>> getSignsByCity(@PathVariable(value = "id") long id) {
+    public ResponseEntity<List<Signal>> getSignsByCity(@PathVariable(value = "id") Integer id) {
         List<Signal> signs = signalService.getSignsByCityId(id);
         return ResponseEntity.status(HttpStatus.OK).body(signs);
     }
